# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Violet Enterprise API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Categories ─────────────────────────────────────────────────────────────
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales transactions
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesPage"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale (POS checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sales/{id}/refund:
    post:
      operationId: refundSale
      tags: [sales]
      summary: Refund a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundInput"
      responses:
        "200":
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # ─── Inventory ──────────────────────────────────────────────────────────────
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"

  # ─── Plans ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    Product:
      type: object
      required: [id, name, sku, price, stock, tenantId]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        price:
          type: number
        costPrice:
          type: ["number", "null"]
        stock:
          type: number
        minStock:
          type: number
        categoryId:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        tenantId:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, sku, price]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        sku:
          type: string
        barcode:
          type: string
        price:
          type: number
          minimum: 0
        costPrice:
          type: number
          minimum: 0
        stock:
          type: number
          default: 0
        minStock:
          type: number
          default: 5
        categoryId:
          type: string
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        sku:
          type: string
        barcode:
          type: string
        price:
          type: number
          minimum: 0
        costPrice:
          type: number
          minimum: 0
        stock:
          type: number
        minStock:
          type: number
        categoryId:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
    ProductsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
        page:
          type: number
        limit:
          type: number

    # ── Customer ───────────────────────────────────────────────────────────────
...
    SaleItem:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        discount:
          type: number
        totalPrice:
          type: number
    Sale:
      type: object
      required: [id, totalAmount, paymentMethod, status, tenantId, createdAt, items]
      properties:
        id:
          type: string
        receiptNumber:
          type: string
        customerId:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        subtotal:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, bank_transfer, store_credit, gift_card, mixed]
        status:
          type: string
          enum: [completed, refunded, partial_refund, voided]
        cashierId:
          type: string
        cashierName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        tenantId:
          type: string
        createdAt:
          type: string
    SaleItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: string
        quantity:
          type: number
          minimum: 1
        unitPrice:
          type: number
        discount:
          type: number
          default: 0
    SaleInput:
      type: object
      required: [items, paymentMethod, totalAmount]
      properties:
        customerId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
          minItems: 1
        paymentMethod:
          type: string
          enum: [cash, card, bank_transfer, store_credit, gift_card, mixed]
        discountAmount:
          type: number
          default: 0
        taxAmount:
          type: number
          default: 0
        totalAmount:
          type: number
        cashTendered:
          type: number
        note:
          type: string
    RefundInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        items:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: string
              quantity:
                type: number
    SalesPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        total:
          type: number
        page:
          type: number
        limit:
          type: number

    # ── Inventory ──────────────────────────────────────────────────────────────
...
    Supplier:
      type: object
      required: [id, name, tenantId]
      properties:
        id:
          type: string
        name:
          type: string
        contactName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tenantId:
          type: string
        createdAt:
          type: string
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string

    # ── Plans & Subscriptions ──────────────────────────────────────────────────
...