# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Violet Enterprise API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  sku: string;
  barcode?: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  costPrice?: number;
  stock?: number;
  minStock?: number;
  categoryId?: string;
  imageUrl?: string;
}